rpi0/requirements.txt
gevent==24.10.3
greenlet==3.1.1
numpy==2.1.3
pynmea2==1.19.0
pyserial==3.5
smbus2==0.5.0
sparkfun-qwiic-icm20948==0.0.1
sparkfun_qwiic_i2c==1.0.0
zope.event==5.0
zope.interface==7.1.1
